test-data/requirements.txt
# Python dependencies for test data generation
kafka-python>=2.0.2
